+++ memote/suite/api.py
@@ -27,6 +27,7 @@ except ImportError:
     import json
 
 import pytest
+from six import iteritems
 
 from memote.suite import TEST_DIRECTORY
 from memote.suite.collect import ResultCollectionPlugin
@@ -71,8 +72,25 @@ def test_model(model, filename=None, results=False, pytest_args=None):
     if filename is not None:
         with open(filename, "w") as file_h:
             LOGGER.info("Writing JSON output '%s'.", filename)
-            json.dump(plugin.results, file_h, sort_keys=True, indent=4,
-                      separators=(",", ": "))
+            try:
+                json.dump(plugin.results, file_h, sort_keys=True, indent=4,
+                          separators=(",", ": "))
+            except TypeError:
+                # Log information to easily find the culprit.
+                json_types = (type(None), int, float, str, list, dict)
+                for mod, functions in iteritems(plugin.results["report"]):
+                    LOGGER.debug("%s:", mod)
+                    for name, annotation in iteritems(functions):
+                        data = annotation.get("data")
+                        try:
+                            for key, value in iteritems(data):
+                                if not isinstance(value, json_types):
+                                    LOGGER.debug(
+                                        "  %s - %s: %s", name, key, type(value))
+                        except AttributeError:
+                            if not isinstance(data, json_types):
+                                LOGGER.debug("  %s: %s", name, type(data))
+
     if results:
         return code, plugin.results
     else: